requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
python-multipart==0.0.6
ironxl==2024.3.0.3
python-dotenv==1.0.0
python-dateutil==2.9.0.post0
psycopg2-binary==2.9.6